New Version Near Sources In core.tcl-lang.org/tcludp.
This is a version which is as near to the official version from core.tcl-lang.org as possible to create 32 and 64 bit extensions for Windows (only Intel / AMD, not ARM) using MSys2 and VSCodium with C/C++ Themes, Makefile Tools and Native Debug extensions.
Folder MSys has been added to hold everything necessary to build the extensions for Windows with VSCodium or pure MSys. To build the version with pure MSys, the environment variables ActualTclPath, TargetType, TargetMachine and PATH should be set before calling make. Values for the different versions (debug, release, 32 or 64 bit architecture) can be found in MSys/.vscode/tasks.json.